requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
imageio
pillow-avif-plugin
redis==4.2.0
celery==5.2.3
dynaconf==3.1.7
marshmallow==3.14.1